+++ library/inspector/src/main/java/io/mehow/laboratory/inspector/ToolbarBinding.kt
@@ -1,20 +1,14 @@
 package io.mehow.laboratory.inspector
 
 import android.view.View
-import android.view.animation.AccelerateDecelerateInterpolator
 import androidx.appcompat.widget.AppCompatEditText
 import androidx.appcompat.widget.AppCompatImageView
-import androidx.constraintlayout.widget.ConstraintLayout
-import androidx.constraintlayout.widget.ConstraintSet
-import androidx.constraintlayout.widget.ConstraintSet.GONE
-import androidx.constraintlayout.widget.ConstraintSet.VISIBLE
+import androidx.constraintlayout.motion.widget.MotionLayout
 import androidx.core.view.isVisible
 import androidx.core.widget.addTextChangedListener
-import androidx.transition.AutoTransition
-import androidx.transition.TransitionManager
-import androidx.transition.TransitionSet.ORDERING_TOGETHER
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
-import com.google.android.material.textview.MaterialTextView
+import io.mehow.laboratory.inspector.SearchMode.Active
+import io.mehow.laboratory.inspector.SearchMode.Idle
 import io.mehow.laboratory.inspector.SearchViewModel.Event.ToggleSearchMode
 import io.mehow.laboratory.inspector.SearchViewModel.Event.UpdateQuery
 import io.mehow.laboratory.inspector.SearchViewModel.SearchUiModel
@@ -24,11 +18,9 @@ internal class ToolbarBinding(
   private val onSearchEventsListener: (SearchViewModel.Event) -> Unit,
   private val onResetEventsListener: () -> Unit,
 ) {
-  private val toolbar = view.findViewById<ConstraintLayout>(R.id.io_mehow_laboratory_toolbar)
-  private val title = view.findViewById<MaterialTextView>(R.id.io_mehow_laboratory_title)
+  private val toolbar = view.findViewById<MotionLayout>(R.id.io_mehow_laboratory_toolbar)
   private val searchQuery = view.findViewById<AppCompatEditText>(R.id.io_mehow_laboratory_feature_query)
-  private val closeSearch = view.findViewById<AppCompatImageView>(R.id.io_mehow_laboratory_close_search)
-  private val showSearch = view.findViewById<AppCompatImageView>(R.id.io_mehow_laboratory_show_search)
+  private val toggleSearch = view.findViewById<KyrieImageView>(R.id.io_mehow_laboratory_toggle_search)
   private val clearQuery = view.findViewById<AppCompatImageView>(R.id.io_mehow_laboratory_clear_query)
   private val resetFeatures = view.findViewById<AppCompatImageView>(R.id.io_mehow_laboratory_reset_features)
 
@@ -47,50 +39,28 @@ internal class ToolbarBinding(
       oldText = query
       onSearchEventsListener(UpdateQuery(query))
     }
-    closeSearch.setOnClickListener { onSearchEventsListener(ToggleSearchMode) }
-    showSearch.setOnClickListener { onSearchEventsListener(ToggleSearchMode) }
+
+    toolbar.setTransition(R.id.io_mehow_laboratory_search_transition)
+    toggleSearch.setOnClickListener { onSearchEventsListener(ToggleSearchMode) }
     clearQuery.setOnClickListener { searchQuery.setText("") }
     resetFeatures.setOnClickListener { resetFeaturesDialog.show() }
   }
 
   fun render(uiModel: SearchUiModel) {
-    TransitionManager.beginDelayedTransition(toolbar, transition)
-    visibilityConstraints(uiModel.showSearch).applyTo(toolbar)
-    clearQuery.isVisible = uiModel.showSearch && uiModel.query.isNotEmpty()
+    uiModel.mode.applyTransition(toolbar)
+    clearQuery.springVisibility(isVisible = uiModel.showSearch && uiModel.query.isNotEmpty())
 
     if (uiModel.showSearch) {
       searchQuery.focusAndShowKeyboard()
     } else {
-      searchQuery.setText("")
       searchQuery.hideKeyboard()
+      searchQuery.setText("")
     }
   }
 
-  private val transition get() = AutoTransition().apply {
-    ordering = ORDERING_TOGETHER
-    duration = 350L
-    interpolator = AccelerateDecelerateInterpolator()
-  }
-
-  private fun visibilityConstraints(showSearch: Boolean) = if (showSearch) {
-    activeConstraints
-  } else {
-    idleConstraints
-  }
-
-  private val idleConstraints = ConstraintSet().apply {
-    clone(toolbar)
-    setVisibility(title.id, VISIBLE)
-    setVisibility(showSearch.id, VISIBLE)
-    setVisibility(closeSearch.id, GONE)
-    setVisibility(searchQuery.id, GONE)
-  }
-
-  private val activeConstraints = ConstraintSet().apply {
-    clone(toolbar)
-    setVisibility(title.id, GONE)
-    setVisibility(showSearch.id, GONE)
-    setVisibility(closeSearch.id, VISIBLE)
-    setVisibility(searchQuery.id, VISIBLE)
-  }
+  private val SearchMode.applyTransition
+    get() = when (this) {
+      Idle -> MotionLayout::transitionToStart
+      Active -> MotionLayout::transitionToEnd
+    }
 }